refactor(gui): split im settings components
